Parallelize export calls
We expect a large number of metrics to be pushed through our
exporters, which is why we're replacing the synchronous export
queue with a goroutine pool.

We're going to leverage the "pond"[1] library, which seems very
promissing (good adoption rate, well documented, simple API and
implementation, said to be efficient).

[1] https://github.com/alitto/pond